Further iterates on the GradeFormula calculation and adds a powerschool API endpoint for psm_reportingterms.

+++ database/createTableDDL.sql
@@ -184,7 +184,7 @@ CREATE TABLE `scholar_warehouse`.`section` (
   `grade_formula` VARCHAR(1024) NULL COMMENT 'The section grading formula as a string',
   `course_fk` BIGINT UNSIGNED NOT NULL COMMENT 'The foreign key to the school table',
   `term_fk` BIGINT UNSIGNED NOT NULL COMMENT 'The foreign key to the term table',
-  `number_of_terms` BIGINT UNSIGNED NULL COMMENT 'The number of terms that the section spans',
+  `number_of_terms` INTEGER UNSIGNED NULL COMMENT 'The number of terms that the section spans',
   `section_source_system_id` VARCHAR(256) NULL UNIQUE COMMENT 'The source system from which the entity was imported - the id from that system',
   PRIMARY KEY (`section_id`),
   CONSTRAINT `fk_course$section`
